Global option for setting a transparent background on dark themes
Global option for setting a transparent background on dark themes,
useful if you like to set your terminal window with some degree of
transparency. The boolean variable exposed to the user is:

transparent_dark_background

If is set to true, the transparent background will be applied to the
dark theme, if is set to false, or not setted at all, then the
background will be set to the default solid color.