feat: implement support for monitoring D-Bus signals
Implements support for using `dbus-monitor`, which allows direct monitoring of
signals, which are send when the settings changes.

This reduces the system load, as continuous polling is no longer required, as
well as improving the responsiveness, as the theme is now updated immediately,
instead of waiting for the next polling-cycle to complete.

If `dbus-monitor` is not available, the polling method will be used as a
fallback.